# services/agentatlas/api/openapi/atlas-agent.yaml
openapi: 3.0.3
info:
  title: AgentAtlas Agent Control Plane API
  version: 1.0.0
  description: >
    Agent-first control plane for admins and knowledge owners: Knowledge Agent
    runs, workflow drafts and publication, dream policies, and artifact
    processing jobs. Drafts may be created by the agent; publication always
    requires an explicit human confirmation. Every call authenticates with an
    AgentNexus-issued ticket.
security:
  - NexusTicket: []
paths:
...
  /api/assessments/{id}:
    get:
      security: [{AtlasSession: []}]
      operationId: getManagerAssessmentDetail
      description: Returns the authorized manager assessment detail (score/level and manager notes) only for a subject within the manager's exact org hierarchy scope and under a policy they own. It exposes no HR-action capability.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200': {description: The authorized manager assessment detail, content: {application/json: {schema: {type: object}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {description: The manager is outside the subject's reporting line or does not own the governing policy.}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {description: The assessment surface is not configured.}
...
components:
  securitySchemes:
    NexusTicket:
      type: apiKey
      in: header
      name: X-Nexus-Ticket
    AtlasSession:
      type: apiKey
      in: cookie
      name: atlas_session
  responses:
...
    Unauthorized:
      description: Missing or invalid ticket (fail closed)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found in caller-visible scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...